backend/requirements.txt
# Core
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.9.0
django-environ==0.11.2

# Celery + Redis
celery==5.4.0
redis==5.3.1

# Database driver (PostgreSQL)
psycopg2-binary==2.9.9
